split analysis module
